meson: Don't specify warning level manually.
Since we already specify a `warning_level` in the project settings, it
is redundant to manually specify it as compiler parameters.

This patch resolves the following meson warning:
```
meson.build:25: WARNING: Consider using the built-in warning_level option instead of using "-Wall".
meson.build:25: WARNING: Consider using the built-in warning_level option instead of using "-Wextra".
```