meson: Raise default warning level.
Increases the default warning level to 3, which includes pedantic
errors. Disable them on C++, which is the way it was done before.

To resolve most pedantic warnings on C++ either the compiler standard
would need to be increased to c++2a or the remaining copositor backends
would require porting to plain C. One of each or both should happen in
the future.