d/wivrn: Add network streaming driver
WiVRn stands for WIreless Virtual Reality n (n stands for nothing)

The driver relies on a standalone headset running wivrn application, the
application multicasts an announce packet, monado driver listens for
this packet and then initiates a connection.

When a connection is established:
Monado driver creates a compositor target, backed by a pseudo-swapchain
which encodes frames in h264/h265.
The application sends tracking data regularly, which is stored on monado
and served when pose, view or input are requested.

Co-authored-by: Guillaume Meunier <guillaume.meunier@centraliens.net>